Pass the object into the modifier call
Pass the object currently being serialized into the modifier call so that the modifier callable has access to the object's context. This will allow the modifier to make modifications to a value based on the current state of the object being modified. For example it allows optional computation of a duration from a start and end date if present.